# src/main/resources/openapi/causa-api.yaml
openapi: 3.0.3
info:
  title: Causa API
  version: 0.2.0
  description: API for issue tracking and phase workflows.
servers:
  - url: /
paths:
...
  /projects:
    get:
      tags:
        - Projects
      summary: List projects
      operationId: listProjects
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: orgId
          in: query
          required: false
          schema:
            type: string
        - name: ownerId
          in: query
          required: false
          schema:
            type: string
        - name: teamId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectSummary'
  /projects/facets:
    get:
      tags:
        - Projects
      summary: List project facets
      operationId: getProjectFacets
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: orgId
          in: query
          required: false
          schema:
            type: string
        - name: ownerId
          in: query
          required: false
          schema:
            type: string
        - name: teamId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Project facet response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectFacetResponse'
  /projects/{projectId}/owner:
    patch:
      tags:
        - Projects
      summary: Assign project owner
      operationId: assignProjectOwner
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignProjectOwnerRequest'
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSummary'
...
components:
  schemas:
...
    ProjectSummary:
      type: object
      required:
        - id
        - key
        - name
        - orgId
        - teamId
      properties:
        id:
          type: string
        key:
          type: string
        name:
          type: string
        orgId:
          type: string
        teamId:
          type: string
        ownerId:
          type: string
          nullable: true
        issueStatusCounts:
          type: object
          additionalProperties:
            type: integer
            format: int64
        phaseStatusCounts:
          type: object
          additionalProperties:
            type: integer
            format: int64
        phaseStatusByIssueStatus:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: integer
              format: int64
    ProjectFacetOption:
      type: object
      required:
        - id
        - count
      properties:
        id:
          type: string
        count:
          type: integer
          format: int64
    ProjectFacetResponse:
      type: object
      required:
        - owners
        - teams
      properties:
        owners:
          type: array
          items:
            $ref: '#/components/schemas/ProjectFacetOption'
        teams:
          type: array
          items:
            $ref: '#/components/schemas/ProjectFacetOption'
    AssignProjectOwnerRequest:
      type: object
      properties:
        ownerId:
          type: string
          nullable: true
...